requirements.txt
# requirements.txt
discord.py==2.3.2
pillow==10.2.0